+++ tests/test_vams_wrap.py
@@ -52,4 +52,5 @@ module myblk_wrap (
         .f(f)
     );
 
-endmodule'''
+endmodule
+'''
